+++ minecraft/src/macohi/MacohiVersion.java
@@ -1,5 +1,5 @@
 package macohi;
 
 public class MacohiVersion {
-	public static String version = "Minecraft Alpha 1.2.6 (m1.2)";
+	public static String version = "Minecraft Alpha 1.2.6 (m1.2_01)";
 }